requirements.txt
pyserial>=3.5
openpyxl>=3.1
fpdf2>=2.7
matplotlib>=3.8
Pillow>=10.0
